+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.1.0 TBA
 
 * increased performance of view helper
+* added configuration for view helper
 
 ## Version 1.0.0 (2014-06-10)